sandbox - move workspace editing service to electron-sandbox
