+++ src/vs/workbench/services/workspaces/electron-sandbox/workspaceEditingService.ts
@@ -13,7 +13,7 @@ import { WorkspaceService } from 'vs/workbench/services/configuration/browser/co
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IBackupFileService } from 'vs/workbench/services/backup/common/backup';
-import { toBackupWorkspaceResource } from 'vs/workbench/services/backup/electron-browser/backup';
+import { toBackupWorkspaceResource } from 'vs/workbench/services/backup/electron-sandbox/backup';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { basename } from 'vs/base/common/resources';
 import { INotificationService, Severity } from 'vs/platform/notification/common/notification';
